fix naming, make get_settings use get_editable_fields method
