+++ backend/oqtopus_cloud/user/routers/users.py
@@ -3,7 +3,7 @@ from os import environ
 from datetime import datetime, timedelta, timezone
 from fastapi import APIRouter, Depends, Request as Event, Body, status
 import boto3
-from oqtopus_cloud.user.schemas.settings import EditableField
+from oqtopus_cloud.user.common.settings import get_editable_fields
 import pytz
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -257,17 +257,6 @@ def retrieve_user_login_history(cognito_id: str, region: str) -> list[LoginEvent
             )
 
     return event_list
-
-
-def get_editable_fields() -> list[EditableField]:
-    try:
-        editable_fields = json.loads(environ.get("EDITABLE_FIELDS", "[]"))
-        if not isinstance(editable_fields, list):
-            editable_fields = []
-
-        return [EditableField(v) for v in editable_fields]
-    except Exception:
-        return []
 
 
 def localize(dt: datetime | None) -> datetime | None:
